+++ server/src/test/java/com/inventage/portal/gateway/proxy/config/dynamic/DynamicConfigurationTest.java
@@ -448,8 +448,26 @@ public class DynamicConfigurationTest {
                                                         .put(DynamicConfiguration.SERVICE_SERVER_HOST, "localhost")
                                                         .put(DynamicConfiguration.SERVICE_SERVER_PORT, 1234))))));
 
-
-
+        JsonObject controlApiMiddlewareWithSessionReset = new JsonObject().put(DynamicConfiguration.HTTP,
+                new JsonObject()
+                        .put(DynamicConfiguration.ROUTERS,
+                                new JsonArray().add(new JsonObject().put(DynamicConfiguration.ROUTER_NAME, "routerFoo")
+                                        .put(DynamicConfiguration.MIDDLEWARES, new JsonArray().add("middlewareFoo"))
+                                        .put(DynamicConfiguration.ROUTER_SERVICE, "serviceFoo")))
+                        .put(DynamicConfiguration.MIDDLEWARES, new JsonArray().add(new JsonObject()
+                                .put(DynamicConfiguration.MIDDLEWARE_NAME, "middlewareFoo")
+                                .put(DynamicConfiguration.MIDDLEWARE_TYPE,
+                                        DynamicConfiguration.MIDDLEWARE_CONTROL_API)
+                                .put(DynamicConfiguration.MIDDLEWARE_OPTIONS, new JsonObject()
+                                        .put(DynamicConfiguration.MIDDLEWARE_CONTROL_API_ACTION, "SESSION_RESET")
+                                )))
+                        .put(DynamicConfiguration.SERVICES,
+                                new JsonArray()
+                                        .add(new JsonObject().put(DynamicConfiguration.SERVICE_NAME, "serviceFoo").put(
+                                                DynamicConfiguration.SERVICE_SERVERS,
+                                                new JsonArray().add(new JsonObject()
+                                                        .put(DynamicConfiguration.SERVICE_SERVER_HOST, "localhost")
+                                                        .put(DynamicConfiguration.SERVICE_SERVER_PORT, 1234))))));
 
         // the sole purpose of the following variable are to improve readability
         boolean expectedTrue = true;
@@ -529,6 +547,7 @@ public class DynamicConfigurationTest {
 
                 // controlapi middleware
                 Arguments.of("accept control api with 'SESSION_TERMINATE' action middleware", controlApiMiddlewareWithSessionTermination, complete, expectedTrue),
+                Arguments.of("accept control api with 'SESSION_RESET' action middleware", controlApiMiddlewareWithSessionReset, complete, expectedTrue),
 
                 // services
                 Arguments.of("reject null services", nullHttpServices, complete, expectedFalse),