+++ server/src/test/java/com/inventage/portal/gateway/proxy/middleware/MiddlewareServer.java
@@ -2,10 +2,7 @@ package com.inventage.portal.gateway.proxy.middleware;
 
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
-import io.vertx.core.http.HttpClientResponse;
-import io.vertx.core.http.HttpMethod;
-import io.vertx.core.http.HttpServer;
-import io.vertx.core.http.RequestOptions;
+import io.vertx.core.http.*;
 import io.vertx.junit5.VertxTestContext;
 
 public class MiddlewareServer {
@@ -13,19 +10,26 @@ public class MiddlewareServer {
     private final Vertx vertx;
     private HttpServer httpServer;
     private final int port;
+    private final String host;
 
-    public MiddlewareServer(Vertx vertx, HttpServer httpServer, int port) {
+    public MiddlewareServer(Vertx vertx, HttpServer httpServer, int port, String host) {
         this.port = port;
         this.httpServer = httpServer;
         this.vertx = vertx;
+        this.host = host;
     }
 
-    public void doRequest(VertxTestContext testCtx, RequestOptions reqOpts, Handler<HttpClientResponse> responseHandler) {
-        reqOpts.setHost("localhost").setPort(port).setURI("/").setMethod(HttpMethod.GET);
-        vertx.createHttpClient().request(reqOpts).compose(req -> req.send()).onComplete(testCtx.succeeding(resp -> {
-            responseHandler.handle(resp);
-            httpServer.close();
+    public void incomingRequest(VertxTestContext testCtx, RequestOptions reqOpts, Handler<HttpClientResponse> responseHandler) {
+        reqOpts.setHost(host).setPort(port).setURI("/").setMethod(HttpMethod.GET);
+        httpServer.listen(port, host).onComplete(testCtx.succeeding(p -> {
+            createHttpClientWithRequestOptionsAndResponseHandler(testCtx, reqOpts, responseHandler);
         }));
     }
 
+    private void createHttpClientWithRequestOptionsAndResponseHandler(VertxTestContext testCtx, RequestOptions reqOpts,
+                                                                      Handler<HttpClientResponse> responseHandler) {
+        vertx.createHttpClient().request(reqOpts).compose(HttpClientRequest::send).onComplete(testCtx.succeeding(resp -> {
+            responseHandler.handle(resp);
+        }));
+    }
 }