+++ server/src/test/java/com/inventage/portal/gateway/proxy/middleware/cors/CorsMiddlewareTest.java
@@ -9,26 +9,28 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-import static com.inventage.portal.gateway.proxy.middleware.MiddlewareServerBuilder.httpServer;
+import static com.inventage.portal.gateway.proxy.middleware.MiddlewareServerBuilder.portalGateway;
 import static io.vertx.core.http.HttpHeaders.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @ExtendWith(VertxExtension.class)
 public class CorsMiddlewareTest {
 
+    private static final String host = "localhost";
     private int port;
 
     @BeforeEach
-    public void setup() throws Exception {
+    public void setup() {
         port = TestUtils.findFreePort();
     }
 
     @Test
     public void test_GET_no_origin(Vertx vertx, VertxTestContext testCtx) throws InterruptedException {
         // given
-        httpServer(vertx, port).withCorsMiddleware("http://portal.minikube")
+        portalGateway(vertx, host, port).withCorsMiddleware("http://portal.minikube")
+                .build()
                 // when
-                .doRequest(testCtx, new RequestOptions(), (resp) -> {
+                .incomingRequest(testCtx, new RequestOptions(), (resp) -> {
                     // then
                     assertEquals(200, resp.statusCode(), "unexpected status code");
                     assertEquals("origin", resp.getHeader(VARY));
@@ -39,9 +41,10 @@ public class CorsMiddlewareTest {
     @Test
     public void test_GET_origin_allowed(Vertx vertx, VertxTestContext testCtx) throws InterruptedException {
         // given
-        httpServer(vertx, port).withCorsMiddleware("http://portal.minikube")
+        portalGateway(vertx, host, port).withCorsMiddleware("http://portal.minikube")
+                .build()
                 // when
-                .doRequest(testCtx, new RequestOptions().addHeader(ORIGIN, "http://portal.minikube"), (resp) -> {
+                .incomingRequest(testCtx, new RequestOptions().addHeader(ORIGIN, "http://portal.minikube"), (resp) -> {
                     // then
                     assertEquals(200, resp.statusCode(), "unexpected status code");
                     assertEquals("http://portal.minikube", resp.getHeader(ACCESS_CONTROL_ALLOW_ORIGIN));
@@ -53,9 +56,10 @@ public class CorsMiddlewareTest {
     @Test
     public void test_GET_all_allowed(Vertx vertx, VertxTestContext testCtx) throws InterruptedException {
         // given
-        httpServer(vertx, port).withCorsMiddleware("*")
+        portalGateway(vertx, host, port).withCorsMiddleware("*")
+                .build()
                 // when
-                .doRequest(testCtx, new RequestOptions().addHeader(ORIGIN, "http://other.com"), (resp) -> {
+                .incomingRequest(testCtx, new RequestOptions().addHeader(ORIGIN, "http://other.com"), (resp) -> {
                     // then
                     assertEquals(200, resp.statusCode(), "unexpected status code");
                     assertEquals("*", resp.getHeader(ACCESS_CONTROL_ALLOW_ORIGIN));
@@ -67,9 +71,10 @@ public class CorsMiddlewareTest {
     @Test
     public void test_GET_origin_not_allowed(Vertx vertx, VertxTestContext testCtx) throws InterruptedException {
         // given
-        httpServer(vertx, port).withCorsMiddleware("http://portal.minikube")
+        portalGateway(vertx, host, port).withCorsMiddleware("http://portal.minikube")
+                .build()
                 // when
-                .doRequest(testCtx, new RequestOptions().addHeader(ORIGIN, "http://bad.com"), (resp) -> {
+                .incomingRequest(testCtx, new RequestOptions().addHeader(ORIGIN, "http://bad.com"), (resp) -> {
                     // then
                     assertEquals(403, resp.statusCode(), "unexpected status code");
                     testCtx.completeNow();
@@ -79,9 +84,10 @@ public class CorsMiddlewareTest {
     @Test
     public void test_OPTIONS_no_origin2(Vertx vertx, VertxTestContext testCtx) throws InterruptedException {
         // given
-        httpServer(vertx, port).withCorsMiddleware("http://portal.minikube")
+        portalGateway(vertx, host, port).withCorsMiddleware("http://portal.minikube")
+                .build()
                 // when
-                .doRequest(testCtx, new RequestOptions(), (resp) -> {
+                .incomingRequest(testCtx, new RequestOptions(), (resp) -> {
                     // then
                     assertEquals(200, resp.statusCode(), "unexpected status code");
                     assertEquals("origin", resp.getHeader(VARY));